+++ src/main/app.ts
@@ -9,7 +9,7 @@ import * as path from "path";
 import * as favicon from "serve-favicon";
 import { authCheckerUserOnlyFilter } from "./user/auth-checker-user-only-filter";
 import { Helmet, IConfig as HelmetConfig } from "./modules/helmet";
-import { RouterFinder } from "./router/routerFinder";
+import { importAll } from "./import-all/index";
 
 const enableAppInsights = require("./app-insights/app-insights");
 
@@ -75,7 +75,7 @@ if (config.useCSRFProtection === true) {
 
 app.all(/^\/(?!oauth2redirect|health).*$/, authCheckerUserOnlyFilter);
 app.all(/^\/(?!oauth2redirect|health).*$/, serviceFilter);
-app.use("/", RouterFinder.findAll(path.join(__dirname, "routes")));
+app.use("/", importAll(path.join(__dirname, "routes")));
 
 // returning "not found" page for requests with paths not resolved by the router
 app.use((req, res) => {
