+++ src/main/import-all/README.md
@@ -1,0 +1,10 @@
+import-all
+----------
+
+given the path to a directory containing javascript files, returns an array of
+all of their default exports
+
+```javascript
+import importAll from 'import-all';
+const imports = importAll(`${__dirname}/directory`);
+```
